update item
working update item :)

+++ ca1/app/src/main/java/ie/wit/ca1/models/CollectionModel.kt
@@ -2,10 +2,11 @@ package ie.wit.ca1.models
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import kotlin.random.Random
 
 @Parcelize
 data class CollectionModel(
-    var id: Long = 0,
+    var id: Int = 0,
     var title: String = "",
     var genre: String = "",
     var cards: ArrayList<String> = ArrayList()
